Upgrade to Gatsby V2
from https://github.com/CaenCamp/new-website/pull/63
great thanks to @mikaoelitiana
