+++ package.json
@@ -24,26 +24,28 @@
     },
     "dependencies": {
         "axios": "0.18.0",
+        "babel-plugin-styled-components": "^1.8.0",
         "date-fns": "1.29.0",
         "font-awesome": "4.7.0",
-        "gatsby": "1.9.260",
-        "gatsby-link": "1.6.44",
-        "gatsby-plugin-react-helmet": "2.0.11",
+        "gatsby": "^2.0.0",
+        "gatsby-link": "^2.0.0",
+        "gatsby-plugin-react-helmet": "^3.0.0",
         "gatsby-plugin-react-next": "1.0.11",
-        "gatsby-plugin-styled-components": "2.0.11",
-        "gatsby-plugin-typography": "1.7.18",
-        "gatsby-source-filesystem": "1.5.35",
-        "gatsby-transformer-remark": "1.7.41",
+        "gatsby-plugin-styled-components": "^3.0.0",
+        "gatsby-plugin-typography": "^2.0.0",
+        "gatsby-source-filesystem": "^2.0.1",
+        "gatsby-transformer-remark": "^2.1.6",
         "prop-types": "15.6.1",
-        "react": "16.3.2",
-        "react-dom": "16.3.2",
+        "react": "^16.5.2",
+        "react-dom": "^16.5.2",
         "react-fontawesome": "1.6.1",
         "react-helmet": "5.2.0",
         "react-player": "1.5.0",
-        "react-typography": "0.16.13",
+        "react-typography": "^0.16.13",
         "storybook": "1.0.0",
         "storybook-router": "0.3.4",
         "styled-components": "3.2.6",
+        "typography": "^0.16.17",
         "url-search-params-polyfill": "5.0.0"
     },
     "devDependencies": {
@@ -51,6 +53,7 @@
         "@storybook/react": "3.4.4",
         "babel-eslint": "8.2.3",
         "babel-jest": "22.4.3",
+        "babel-loader": "7.1.1",
         "babel-preset-es2015": "6.24.1",
         "babel-preset-react": "6.24.1",
         "enzyme": "3.3.0",
