+++ src/jestPreprocess.js
@@ -1,0 +1,3 @@
+const babelOptions = { presets: ['es2015', 'react'] };
+
+module.exports = require('babel-jest').createTransformer(babelOptions);
