+++ src/pages/coding-caen-camp.js
@@ -1,7 +1,9 @@
+import { graphql } from 'gatsby';
 import { Helmet } from 'react-helmet';
 import React from 'react';
 import styled from 'styled-components';
 
+import Layout from '../components/layout';
 import { Content, SingleColumn } from '../components/Content';
 import CCC from '../components/CodingCaenCamp';
 import { formatGraphContent } from '../utils/formatters';
@@ -19,33 +21,35 @@ export const Thanks = styled.div`
 export default ({ data }) => {
     const cccs = data.cccs.edges.map(camp => formatGraphContent(camp.node));
     return (
-        <div>
-            <Helmet title="CaenCamp: les Coding Caen Camp (CCC)">
-                <meta
-                    name="description"
-                    content="Affutez vos skills aux Coding Caen Camp"
-                />
-            </Helmet>
-            <Content id="dojoContent">
-                <SingleColumn>
-                    <CCC />
-                    {cccs &&
-                        cccs.map(camp => (
-                            <CampListItem key={camp.id} camp={camp} />
-                        ))}
-                    <Thanks>
-                        <p>
-                            Un grand merci à Emmanuelle et Sylvain de{' '}
-                            <a href="https://www.hey-coworking.com/">
-                                HEY! coworking
-                            </a>{' '}
-                            pour avoir accepter de nous accueillir une fois par
-                            mois.
-                        </p>
-                    </Thanks>
-                </SingleColumn>
-            </Content>
-        </div>
+        <Layout>
+            <div>
+                <Helmet title="CaenCamp: les Coding Caen Camp (CCC)">
+                    <meta
+                        name="description"
+                        content="Affutez vos skills aux Coding Caen Camp"
+                    />
+                </Helmet>
+                <Content id="dojoContent">
+                    <SingleColumn>
+                        <CCC />
+                        {cccs &&
+                            cccs.map(camp => (
+                                <CampListItem key={camp.id} camp={camp} />
+                            ))}
+                        <Thanks>
+                            <p>
+                                Un grand merci à Emmanuelle et Sylvain de{' '}
+                                <a href="https://www.hey-coworking.com/">
+                                    HEY! coworking
+                                </a>{' '}
+                                pour avoir accepter de nous accueillir une fois
+                                par mois.
+                            </p>
+                        </Thanks>
+                    </SingleColumn>
+                </Content>
+            </div>
+        </Layout>
     );
 };
