+++ src/pages/welcome-us.js
@@ -2,8 +2,9 @@ import { Helmet } from 'react-helmet';
 import React from 'react';
 
 import { Content, LeftColumn } from '../components/Content';
-import SideMenu from '../components/SideMenu';
 import { formatMeetup } from '../utils/formatters';
+import Layout from '../components/layout';
+import SideMenu from '../components/SideMenu';
 
 export default ({ nextMeetup }) => {
     if (nextMeetup) {
@@ -11,16 +12,21 @@ export default ({ nextMeetup }) => {
     }
 
     return (
-        <div>
-            <Helmet title="CaenCamp: soutenez-nous">
-                <meta name="description" content="Accueillez les CaenCamp" />
-            </Helmet>
-            <Content id="welcomeUsContent">
-                <LeftColumn>
-                    <h1>Accueillez nous</h1>
-                </LeftColumn>
-                <SideMenu meetup={nextMeetup} />
-            </Content>
-        </div>
+        <Layout>
+            <div>
+                <Helmet title="CaenCamp: soutenez-nous">
+                    <meta
+                        name="description"
+                        content="Accueillez les CaenCamp"
+                    />
+                </Helmet>
+                <Content id="welcomeUsContent">
+                    <LeftColumn>
+                        <h1>Accueillez nous</h1>
+                    </LeftColumn>
+                    <SideMenu meetup={nextMeetup} />
+                </Content>
+            </div>
+        </Layout>
     );
 };
