+++ src/templates/dojo.js
@@ -1,12 +1,14 @@
+import { graphql } from 'gatsby';
 import Helmet from 'react-helmet';
-import React from 'react';
 import Link from 'gatsby-link';
+import React from 'react';
 import styled from 'styled-components';
 
 import { formatDojoWithCraftsmen } from '../utils/formatters';
 import { SingleColumn } from '../components/Content';
-import SpeakerTalk from '../components/speakers/SpeakerTalk';
 import Calendar from '../components/talks/Calendar';
+import Layout from '../components/layout';
+import SpeakerTalk from '../components/speakers/SpeakerTalk';
 import Tags from '../components/talks/Tags';
 
 const StyledLink = styled(Link)`
@@ -44,30 +46,32 @@ export const Title = styled.h1`
 export default ({ data }) => {
     const dojo = formatDojoWithCraftsmen(data.rawDojo, data.craftsmen.edges);
     return (
-        <SingleColumn>
-            <Helmet>
-                <title>{dojo.title}</title>
-                <meta name="description" content={dojo.description} />
-                <meta name="keywords" content={`${dojo.tags}`} />
-            </Helmet>
-            <StyledLink to="/coding-dojo">
-                <i className="fa fa-list-alt" aria-hidden="true" /> Retour à la
-                liste
-            </StyledLink>
-            <DojoContainer>
-                <DateAndSpeakers>
-                    <Calendar date={dojo.date} edition={dojo.edition} />
-                    {dojo.craftsmen.map(speaker => (
-                        <SpeakerTalk key={speaker.slug} speaker={speaker} />
-                    ))}
-                </DateAndSpeakers>
-                <Description>
-                    <Title>{dojo.title}</Title>
-                    <Tags tags={dojo.tags} />
-                    <div dangerouslySetInnerHTML={{ __html: dojo.html }} />
-                </Description>
-            </DojoContainer>
-        </SingleColumn>
+        <Layout>
+            <SingleColumn>
+                <Helmet>
+                    <title>{dojo.title}</title>
+                    <meta name="description" content={dojo.description} />
+                    <meta name="keywords" content={`${dojo.tags}`} />
+                </Helmet>
+                <StyledLink to="/coding-dojo">
+                    <i className="fa fa-list-alt" aria-hidden="true" /> Retour à
+                    la liste
+                </StyledLink>
+                <DojoContainer>
+                    <DateAndSpeakers>
+                        <Calendar date={dojo.date} edition={dojo.edition} />
+                        {dojo.craftsmen.map(speaker => (
+                            <SpeakerTalk key={speaker.slug} speaker={speaker} />
+                        ))}
+                    </DateAndSpeakers>
+                    <Description>
+                        <Title>{dojo.title}</Title>
+                        <Tags tags={dojo.tags} />
+                        <div dangerouslySetInnerHTML={{ __html: dojo.html }} />
+                    </Description>
+                </DojoContainer>
+            </SingleColumn>
+        </Layout>
     );
 };
